backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.35.0
sqlalchemy==2.0.38
aiosqlite==0.21.0
asyncpg==0.30.0
pydantic==2.11.0
pydantic-settings==2.8.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.27.0
email-validator==2.3.0
psycopg2-binary==2.9.10
google-auth==2.49.1
google-auth-oauthlib
google-auth-httplib2
python-jose